bancos-dados-big-data/exercicio-mongo-mysql/requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.2
Pillow==9.5.0
pymongo==4.3.3
matplotlib==3.7.1
